Move argument structures for specified subcommands to a new crate
It would be useful to have an independent crate covering solely the OCI
command line front end of youki, to which other backends can then be
written.

To start on that, we create a new "oci-runtime-clap" crate, and move some
subcommand structure definitions into it.  For now we only move the
subcommands which are actually specified in the OCI runtime command line
interface document (v1.0.1):

https://github.com/opencontainers/runtime-tools/blob/master/docs/command-line-interface.md
