+++ crates/liboci-cli/src/lib.rs
@@ -1,0 +1,7 @@
+mod create;
+mod delete;
+mod kill;
+mod start;
+mod state;
+
+pub use {create::Create, delete::Delete, kill::Kill, start::Start, state::State};
