+++ crates/youki/src/commands/create.rs
@@ -1,30 +1,9 @@
 //! Handles the creation of a new container
 use anyhow::Result;
-use clap::Parser;
 use std::path::PathBuf;
 
 use libcontainer::{container::builder::ContainerBuilder, syscall::syscall::create_syscall};
-
-/// Create a container
-#[derive(Parser, Debug)]
-pub struct Create {
-    /// File to write pid of the container created
-    // note that in the end, container is just another process
-    #[clap(short, long)]
-    pid_file: Option<PathBuf>,
-    /// path to the bundle directory, containing config.json and root filesystem
-    #[clap(short, long, default_value = ".")]
-    bundle: PathBuf,
-    /// Unix socket (file) path , which will receive file descriptor of the writing end of the pseudoterminal
-    #[clap(short, long)]
-    console_socket: Option<PathBuf>,
-    /// Pass N additional file descriptors to the container (stdio + $LISTEN_FDS + N in total)
-    #[clap(long, default_value = "0")]
-    preserve_fds: i32,
-    /// name of the container instance to be started
-    #[clap(forbid_empty_values = true, required = true)]
-    pub container_id: String,
-}
+use liboci_cli::Create;
 
 // One thing to note is that in the end, container is just another process in Linux
 // it has specific/different control group, namespace, using which program executing in it
